Separated WFSGeoJSONImportAdapter into 3 files
