+++ Assets/_Functionalities/WFS/Scripts/FeatureType.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2408c28973b874752ac29999e0f8d4f4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
