+++ Assets/_Functionalities/WFS/Scripts/GeoJSONWFS.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e01a12f7c338d46b982498dacc506221
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
